requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
langchain==0.1.0
langgraph==0.0.15
langsmith==0.0.83
boto3==1.34.14
opensearch-py==2.4.2
neo4j==5.16.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
pytest==7.4.4
httpx==0.26.0
langchain-aws==0.1.7
langchain-community==0.0.13
langchain-openai==0.0.2
pandas==2.2.0
numpy==1.26.3
